feat: implement public OCP api to update resources and rooms
