+++ lib/public/Calendar/Room/IManager.php
@@ -10,7 +10,6 @@ namespace OCP\Calendar\Room;
 
 /**
  * @since 14.0.0
- * @deprecated 24.0.0
  */
 interface IManager {
 	/**
@@ -55,4 +54,11 @@ interface IManager {
 	 * @deprecated 24.0.0
 	 */
 	public function clear();
+
+	/**
+	 * Update all rooms from all backends right now.
+	 *
+	 * @since 30.0.0
+	 */
+	public function update(): void;
 }
