+++ tests/lib/Calendar/Room/ManagerTest.php
@@ -12,6 +12,7 @@ namespace Test\Calendar\Room;
 use OC\AppFramework\Bootstrap\Coordinator;
 use OC\AppFramework\Bootstrap\RegistrationContext;
 use OC\AppFramework\Bootstrap\ServiceRegistration;
+use OC\Calendar\ResourcesRoomsUpdater;
 use OC\Calendar\Room\Manager;
 use OCP\Calendar\Room\IBackend;
 use OCP\IServerContainer;
@@ -25,6 +26,9 @@ class ManagerTest extends TestCase {
 	/** @var IServerContainer|MockObject */
 	private $server;
 
+	/** @var ResourcesRoomsUpdater|MockObject */
+	private $resourcesRoomsUpdater;
+
 	/** @var Manager */
 	private $manager;
 
@@ -33,9 +37,12 @@ class ManagerTest extends TestCase {
 
 		$this->coordinator = $this->createMock(Coordinator::class);
 		$this->server = $this->createMock(IServerContainer::class);
+		$this->resourcesRoomsUpdater = $this->createMock(ResourcesRoomsUpdater::class);
+
 		$this->manager = new Manager(
 			$this->coordinator,
 			$this->server,
+			$this->resourcesRoomsUpdater,
 		);
 	}
 
@@ -127,5 +134,12 @@ class ManagerTest extends TestCase {
 		$this->manager->clear();
 
 		self::assertEquals([], $this->manager->getBackends());
+	}
+
+	public function testUpdate(): void {
+		$this->resourcesRoomsUpdater->expects(self::once())
+			->method('updateRooms');
+
+		$this->manager->update();
 	}
 }
